requirements.txt
passlib==1.6.5
urllib3==1.12
wheel==0.26.0
wsgiref==0.1.2
docker-compose==1.5.1
docker-py==1.5.0
dockerpty==0.3.4
ansible==1.9.4
requests==2.9.1
sleekxmpp==1.3.1
ipython==4.1.2
Flask==0.10.1
gunicorn==19.6.0
redis==2.10.5
plant==0.1.2
flask-session==0.2.3
coloredlogs==5.0
Flask-Login==0.2.11
cryptography==1.5
alembic==0.8.7
milieu==0.1.9
PyNaCl==1.0.1
MySQL-python==1.2.5
eyeD3==0.7.9
carbontube==0.0.1
python-dateutil==2.4.2
